# backend/openapi.yaml
openapi: 3.1.0
info:
    title: EduFlow
    version: 0.1.0
    description: |
        EduFlow is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        AnswerBatch:
            description:
                "Body for POST /submissions/{id}/answers \u2014 append/replace\
                \ answers."
            properties:
                answers:
                    items:
                        $ref: "#/components/schemas/AnswerItem"
                    title: Answers
                    type: array
            required:
                - answers
            title: AnswerBatch
            type: object
        AnswerItem:
            properties:
                question_id:
                    title: Question Id
                    type: string
                selected_option_ids:
                    default: []
                    items:
                        type: string
                    title: Selected Option Ids
                    type: array
            required:
                - question_id
            title: AnswerItem
            type: object
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        Page_SubmissionResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/SubmissionResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[SubmissionResponse]
            type: object
...
        SubmissionResponse:
            properties:
                answers:
                    default: []
                    items:
                        $ref: "#/components/schemas/AnswerItem"
                    title: Answers
                    type: array
                assessment_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Assessment Id
                id:
                    title: Id
                    type: string
                pass_fail:
                    anyOf:
                        - type: boolean
                        - type: "null"
                    title: Pass Fail
                percentage:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Percentage
                score:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Score
                student_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Student Id
                submitted_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Submitted At
            required:
                - id
                - student_id
                - assessment_id
                - score
                - percentage
                - submitted_at
            title: SubmissionResponse
            type: object
        SubmissionReviewQuestion:
            properties:
                correct_option_ids:
                    default: []
                    items:
                        type: string
                    title: Correct Option Ids
                    type: array
                correct_option_texts:
                    default: []
                    items:
                        type: string
                    title: Correct Option Texts
                    type: array
                earned_marks:
                    title: Earned Marks
                    type: integer
                is_correct:
                    title: Is Correct
                    type: boolean
                max_marks:
                    title: Max Marks
                    type: integer
                question_id:
                    title: Question Id
                    type: string
                question_text:
                    title: Question Text
                    type: string
                selected_option_ids:
                    default: []
                    items:
                        type: string
                    title: Selected Option Ids
                    type: array
                selected_option_texts:
                    default: []
                    items:
                        type: string
                    title: Selected Option Texts
                    type: array
            required:
                - question_id
                - question_text
                - earned_marks
                - max_marks
                - is_correct
            title: SubmissionReviewQuestion
            type: object
        SubmissionReviewResponse:
            properties:
                assessment_id:
                    title: Assessment Id
                    type: string
                pass_fail:
                    title: Pass Fail
                    type: boolean
                percentage:
                    title: Percentage
                    type: number
                questions:
                    default: []
                    items:
                        $ref: "#/components/schemas/SubmissionReviewQuestion"
                    title: Questions
                    type: array
                score:
                    title: Score
                    type: integer
                student_id:
                    title: Student Id
                    type: string
                submission_id:
                    title: Submission Id
                    type: string
                total_marks:
                    title: Total Marks
                    type: integer
            required:
                - submission_id
                - assessment_id
                - student_id
                - score
                - total_marks
                - percentage
                - pass_fail
            title: SubmissionReviewResponse
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/submissions/assessment/{assessment_id}:
        get:
            operationId: list_by_assessment_api_v1_submissions_assessment__assessment_id__get
            parameters:
                - in: path
                  name: assessment_id
                  required: true
                  schema:
                      title: Assessment Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_SubmissionResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List all submissions for an assessment (staff only)
            tags:
                - submissions
    /api/v1/submissions/student/{student_id}:
        get:
            operationId: list_by_student_api_v1_submissions_student__student_id__get
            parameters:
                - in: path
                  name: student_id
                  required: true
                  schema:
                      title: Student Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_SubmissionResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List all submissions for a student
            tags:
                - submissions
    /api/v1/submissions/{submission_id}/answers:
        post:
            operationId: post_answers_api_v1_submissions__submission_id__answers_post
            parameters:
                - in: path
                  name: submission_id
                  required: true
                  schema:
                      title: Submission Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AnswerBatch"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubmissionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: "Append or replace answers in a submission (student: own only)"
            tags:
                - submissions
    /api/v1/submissions/{submission_id}/grade:
        post:
            operationId: grade_pending_submission_api_v1_submissions__submission_id__grade_post
            parameters:
                - in: path
                  name: submission_id
                  required: true
                  schema:
                      title: Submission Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubmissionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Auto-grade a pending submission (staff only)
            tags:
                - submissions
    /api/v1/submissions/{submission_id}/result:
        get:
            operationId: get_result_api_v1_submissions__submission_id__result_get
            parameters:
                - in: path
                  name: submission_id
                  required: true
                  schema:
                      title: Submission Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubmissionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get graded result of a submission
            tags:
                - submissions
    /api/v1/submissions/{submission_id}/review:
        get:
            operationId: review_submission_api_v1_submissions__submission_id__review_get
            parameters:
                - in: path
                  name: submission_id
                  required: true
                  schema:
                      title: Submission Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubmissionReviewResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Detailed graded submission review for staff
            tags:
                - submissions
...